refactor: split SkillEditor.tsx into custom hooks and models
Addresses remaining PR review feedback: the 900+ line SkillEditor page
component is now decomposed into four focused hooks (useSkillEditorLoad,
useSkillFilePreviewSync, useSkillFileActions, useSkillEditorSubmit),
cutting the page component to routing/orchestration/JSX only. The
skill-file-batch-validation interfaces move to a models/ file, leaving
utils/skill-file-batch-validation.ts to just the validator function.
Also converts the load-state union type to a SkillEditorLoadState string
enum per repo convention.

No behavior change; verified via existing SkillEditor test suite.
